requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langsmith>=0.2.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
pgvector>=0.2.4
alembic>=1.13.1
neo4j>=5.15.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
structlog>=24.1.0
tenacity>=8.2.3
boto3>=1.34.0
mcp>=1.0.0
python-multipart>=0.0.6
langchain-google-genai>=2.0.0
google-auth>=2.27.0
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
networkx>=3.0
uuid
